[REFACTOR] post폴더 디렉토링 변경
